+++ scout/demo/__init__.py
@@ -12,6 +12,7 @@ ped_file = 'demo/643594.ped'
 load_file = 'demo/643594.config.yaml'
 clinical_snv_file = 'demo/643594.clinical.vcf.gz'
 research_snv_file = 'demo/643594.research.vcf.gz'
+customannotation_snv_file = 'demo/customannotations_one.vcf.gz'
 
 ped_path = pkg_resources.resource_filename('scout', ped_file)
 clinical_sv_file = 'demo/643594.clinical.SV.vcf.gz'
@@ -27,6 +28,8 @@ load_path = pkg_resources.resource_filename('scout', load_file)
 clinical_snv_path = pkg_resources.resource_filename('scout', clinical_snv_file)
 clinical_sv_path = pkg_resources.resource_filename('scout', clinical_sv_file)
 clinical_str_path = pkg_resources.resource_filename('scout', clinical_str_file)
+
+customannotation_snv_path = pkg_resources.resource_filename('scout', customannotation_snv_file)
 
 research_snv_path = pkg_resources.resource_filename('scout', research_snv_file)
 research_sv_path = pkg_resources.resource_filename('scout', research_sv_file)